Debounce State
Debounce the button state, not just the pin state. This way debounce() will know if it's pressed or not pressed.

Also, enable a negative ms value so it will only debounce the state on inactive. This way a button press will be recognized instantly. Deactivating the button press will still be subject to the debounce time, preventing switch "noise" from propagating through to the _fsm.